Add url to the Page class
This attribute is used to generate hrefs in HTML page. But it is limited only to
paths ending with html extension. For example, if you have a file:
	test.rhtml
url attribute will be:
	test.html
But, if you have a file:
	test.rhtml.php
url attribute will be:
	test.html.php
At the moment, it's not possible to manage other extesnsions.